Moved column sorting into PerfTable from PerfTracker
